+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog
   - Aliases with variable parts.
   - Unsugared sigils.
   - Calls to `__block__/0`
+  - Aliases with a quoted base module.
 Map updates wand captured map update functions.
 - Fix: Correctly render multi-arity and multi-clause fns that have
   guard clauses.